+++ src/byro/public/views.py
@@ -68,7 +68,7 @@ class MemberView(MemberBaseView):
         context["memberships"] = memberships
         context["member_since"] = {
             "days": int(delta.total_seconds() / (60 * 60 * 24)),
-            "years": round(delta.days / 365, 1),
+            "years": int(round(delta.days / 365, 1)),
             "first": first,
         }
         context["tiles"] = []
